Refactor NumTramite validators testing
